moved to  consensus project and splitted in two behaviors
